feat(#369): use common initialization for all the mojos
