+++ src/main/java/org/eolang/jeo/EoToBytecodeMojo.java
@@ -24,10 +24,13 @@
 package org.eolang.jeo;
 
 import java.io.File;
+import org.apache.maven.artifact.DependencyResolutionRequiredException;
 import org.apache.maven.plugin.AbstractMojo;
+import org.apache.maven.plugin.MojoExecutionException;
 import org.apache.maven.plugins.annotations.LifecyclePhase;
 import org.apache.maven.plugins.annotations.Mojo;
 import org.apache.maven.plugins.annotations.Parameter;
+import org.apache.maven.project.MavenProject;
 import org.eolang.jeo.improvement.ImprovementBytecodeFootprint;
 
 /**
@@ -35,17 +38,26 @@ import org.eolang.jeo.improvement.ImprovementBytecodeFootprint;
  * The mojo that converts EO to bytecode only.
  * It does not apply any improvements.
  *
- * @since 0.1.0
  * @todo #59:90min Duplication between EoToBytecode and Optimization.
- *  Both classes have the same code. We have to extract the common code to the separate class.
- *  The class should be able to convert EO to bytecode and save result to the target folder.
- *  When it is done, remove that puzzle.
+ * Both classes have the same code. We have to extract the common code to the separate class.
+ * The class should be able to convert EO to bytecode and save result to the target folder.
+ * When it is done, remove that puzzle.
+ * @since 0.1.0
  */
 @Mojo(name = "eo-to-bytecode", defaultPhase = LifecyclePhase.PROCESS_CLASSES)
 public final class EoToBytecodeMojo extends AbstractMojo {
 
     /**
+     * Maven project.
+     *
+     * @since 0.2
+     */
+    @Parameter(defaultValue = "${project}", required = true, readonly = true)
+    private MavenProject project;
+
+    /**
      * Project compiled classes.
+     *
      * @since 0.1.0
      */
     @Parameter(defaultValue = "${project.build.outputDirectory}")
@@ -53,14 +65,20 @@ public final class EoToBytecodeMojo extends AbstractMojo {
 
     /**
      * Project default target directory.
+     *
      * @since 0.1.0
      */
     @Parameter(defaultValue = "${project.build.directory}/generated-sources")
     private File generated;
 
     @Override
-    public void execute() {
-        new ImprovementBytecodeFootprint(this.classes.toPath())
-            .apply(new EoRepresentations(this.generated.toPath()).objects());
+    public void execute() throws MojoExecutionException {
+        try {
+            new PluginStartup(this.project).init();
+            new ImprovementBytecodeFootprint(this.classes.toPath())
+                .apply(new EoRepresentations(this.generated.toPath()).objects());
+        } catch (DependencyResolutionRequiredException exception) {
+            throw new MojoExecutionException(exception);
+        }
     }
 }
