+++ src/main/java/org/eolang/jeo/JeoMojo.java
@@ -79,7 +79,7 @@ public final class JeoMojo extends AbstractMojo {
      */
     public void execute() throws MojoExecutionException {
         try {
-            this.initClassloader();
+            new PluginStartup(this.project).init();
             new Optimization(
                 this.classes.toPath(),
                 new ImprovementSet(
@@ -92,46 +92,6 @@ public final class JeoMojo extends AbstractMojo {
         } catch (final IllegalStateException | IOException
             | DependencyResolutionRequiredException exception) {
             throw new MojoExecutionException(exception);
-        }
-    }
-
-    /**
-     * Initialize classloader.
-     * This method is important to load classes that were compiled on the previous maven
-     * phases. Since the jeo plugin works on the 'process-classes' phase, it might
-     * see classes that were compiled on the 'compile' phase.
-     * We need to have all these classes in the classpath to be able to load them during
-     * the transformation phase.
-     * We need this to solve the problem with computing maxs in ASM library:
-     * - https://gitlab.ow2.org/asm/asm/-/issues/317918
-     * - https://stackoverflow.com/questions/11292701/error-while-instrumenting-class-files-asm-classwriter-getcommonsuperclass
-     * @throws DependencyResolutionRequiredException If a problem happened during loading classes.
-     */
-    private void initClassloader() throws DependencyResolutionRequiredException {
-        Thread.currentThread().setContextClassLoader(
-            new URLClassLoader(
-                this.project.getRuntimeClasspathElements()
-                    .stream()
-                    .map(File::new)
-                    .map(JeoMojo::url).toArray(URL[]::new),
-                Thread.currentThread().getContextClassLoader()
-            )
-        );
-    }
-
-    /**
-     * Convert file to URL.
-     * @param file File.
-     * @return URL.
-     */
-    private static URL url(final File file) {
-        try {
-            return file.toURI().toURL();
-        } catch (final MalformedURLException exception) {
-            throw new IllegalStateException(
-                String.format("Can't convert file %s to URL", file),
-                exception
-            );
         }
     }
 }
